requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
streamlit
pillow
matplotlib
numpy
dvc